use onboard computer status message instead
