+++ app/telemetry/mavlinkcommand.cpp
@@ -1,0 +1,3 @@
+#include "mavlinkcommand.h"
+
+
