+++ lib/c_library_v2_openhd/ardupilotmega/ardupilotmega.h
@@ -10,7 +10,7 @@
     #error Wrong include order: MAVLINK_ARDUPILOTMEGA.H MUST NOT BE DIRECTLY USED. Include mavlink.h from the same directory instead or set ALL AND EVERY defines from MAVLINK.H manually accordingly, including the #define MAVLINK_H call.
 #endif
 
-#define MAVLINK_ARDUPILOTMEGA_XML_HASH 2442740817897951245
+#define MAVLINK_ARDUPILOTMEGA_XML_HASH 977177204667597639
 
 #ifdef __cplusplus
 extern "C" {
