+++ lib/c_library_v2_openhd/ardupilotmega/mavlink.h
@@ -6,7 +6,7 @@
 #ifndef MAVLINK_H
 #define MAVLINK_H
 
-#define MAVLINK_PRIMARY_XML_HASH 2442740817897951245
+#define MAVLINK_PRIMARY_XML_HASH 977177204667597639
 
 #ifndef MAVLINK_STX
 #define MAVLINK_STX 253
