+++ lib/c_library_v2_openhd/common/common.h
@@ -10,7 +10,7 @@
     #error Wrong include order: MAVLINK_COMMON.H MUST NOT BE DIRECTLY USED. Include mavlink.h from the same directory instead or set ALL AND EVERY defines from MAVLINK.H manually accordingly, including the #define MAVLINK_H call.
 #endif
 
-#define MAVLINK_COMMON_XML_HASH 7631128334975960780
+#define MAVLINK_COMMON_XML_HASH -678853511729702766
 
 #ifdef __cplusplus
 extern "C" {
