+++ lib/c_library_v2_openhd/common/mavlink.h
@@ -6,7 +6,7 @@
 #ifndef MAVLINK_H
 #define MAVLINK_H
 
-#define MAVLINK_PRIMARY_XML_HASH 7631128334975960780
+#define MAVLINK_PRIMARY_XML_HASH -678853511729702766
 
 #ifndef MAVLINK_STX
 #define MAVLINK_STX 253
