+++ lib/c_library_v2_openhd/icarous/mavlink.h
@@ -6,7 +6,7 @@
 #ifndef MAVLINK_H
 #define MAVLINK_H
 
-#define MAVLINK_PRIMARY_XML_HASH 736758955976123022
+#define MAVLINK_PRIMARY_XML_HASH -3759025632668499863
 
 #ifndef MAVLINK_STX
 #define MAVLINK_STX 253
