+++ lib/c_library_v2_openhd/minimal/mavlink.h
@@ -6,7 +6,7 @@
 #ifndef MAVLINK_H
 #define MAVLINK_H
 
-#define MAVLINK_PRIMARY_XML_HASH 7508872371248599562
+#define MAVLINK_PRIMARY_XML_HASH 3813953899913856194
 
 #ifndef MAVLINK_STX
 #define MAVLINK_STX 253
