+++ lib/c_library_v2_openhd/minimal/minimal.h
@@ -10,7 +10,7 @@
     #error Wrong include order: MAVLINK_MINIMAL.H MUST NOT BE DIRECTLY USED. Include mavlink.h from the same directory instead or set ALL AND EVERY defines from MAVLINK.H manually accordingly, including the #define MAVLINK_H call.
 #endif
 
-#define MAVLINK_MINIMAL_XML_HASH 7508872371248599562
+#define MAVLINK_MINIMAL_XML_HASH 3813953899913856194
 
 #ifdef __cplusplus
 extern "C" {
