+++ lib/c_library_v2_openhd/openhd/mavlink.h
@@ -6,7 +6,7 @@
 #ifndef MAVLINK_H
 #define MAVLINK_H
 
-#define MAVLINK_PRIMARY_XML_HASH 4271000980559245404
+#define MAVLINK_PRIMARY_XML_HASH 4104083464867081584
 
 #ifndef MAVLINK_STX
 #define MAVLINK_STX 253
