+++ lib/c_library_v2_openhd/openhd/openhd.h
@@ -10,7 +10,7 @@
     #error Wrong include order: MAVLINK_OPENHD.H MUST NOT BE DIRECTLY USED. Include mavlink.h from the same directory instead or set ALL AND EVERY defines from MAVLINK.H manually accordingly, including the #define MAVLINK_H call.
 #endif
 
-#define MAVLINK_OPENHD_XML_HASH 4271000980559245404
+#define MAVLINK_OPENHD_XML_HASH 4104083464867081584
 
 #ifdef __cplusplus
 extern "C" {
