+++ lib/c_library_v2_openhd/uAvionix/mavlink.h
@@ -6,7 +6,7 @@
 #ifndef MAVLINK_H
 #define MAVLINK_H
 
-#define MAVLINK_PRIMARY_XML_HASH 4650369192959294724
+#define MAVLINK_PRIMARY_XML_HASH -2350654451775700645
 
 #ifndef MAVLINK_STX
 #define MAVLINK_STX 253
