+++ lib/c_library_v2_openhd/uAvionix/uAvionix.h
@@ -10,7 +10,7 @@
     #error Wrong include order: MAVLINK_UAVIONIX.H MUST NOT BE DIRECTLY USED. Include mavlink.h from the same directory instead or set ALL AND EVERY defines from MAVLINK.H manually accordingly, including the #define MAVLINK_H call.
 #endif
 
-#define MAVLINK_UAVIONIX_XML_HASH 4650369192959294724
+#define MAVLINK_UAVIONIX_XML_HASH -2350654451775700645
 
 #ifdef __cplusplus
 extern "C" {
